% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodiversityhorizons-package.R
\docType{package}
\name{biodiversityhorizons-package}
\alias{biodiversityhorizons}
\alias{biodiversityhorizons-package}
\title{biodiversityhorizons: Biodiversity Horizons Package}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ishika Khandelwal \email{ishikakhandelwal02@gmail.com}

}
\keyword{internal}
