% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{exposure}
\alias{exposure}
\title{Calculate species exposure to climate changes}
\usage{
exposure(data, species_range, climate_data, niche)
}
\arguments{
\item{data}{Data for a single species}

\item{species_range}{List of grid cell IDs for each species}

\item{climate_data}{Data frame of climate data by grid cell}

\item{niche}{Niche limits for each species}
}
\value{
A data frame with exposure data
}
\description{
Calculate species exposure to climate changes
}
