% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_times.R
\name{exposure_times}
\alias{exposure_times}
\title{Calculate exposure times for each species}
\usage{
exposure_times(data, original_state, consecutive_elements)
}
\arguments{
\item{data}{A row of exposure data}

\item{original_state}{Initial exposure state}

\item{consecutive_elements}{Minimum consecutive years for state change}
}
\value{
A tibble with exposure and de-exposure times
}
\description{
Calculate exposure times for each species
}
