% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_climate_data.R
\name{extract_climate_data}
\alias{extract_climate_data}
\title{Extract climate data for each grid cell}
\usage{
extract_climate_data(climate_data, grid)
}
\arguments{
\item{climate_data}{Raster data of climate variables}

\item{grid}{Grid data frame}
}
\value{
A tibble with climate data by grid cell
}
\description{
Extract climate data for each grid cell
}
