% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_niche_limits.R
\name{get_niche_limits}
\alias{get_niche_limits}
\title{Calculate thermal niche limits for each species}
\usage{
get_niche_limits(species_ranges, climate_df)
}
\arguments{
\item{species_ranges}{List of grid cell IDs for each species}

\item{climate_df}{Data frame of climate data by grid cell}
}
\value{
A tibble with upper and lower niche limits
}
\description{
Calculate thermal niche limits for each species
}
