% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_range.R
\name{prepare_range}
\alias{prepare_range}
\title{Prepare range data to match the grid}
\usage{
prepare_range(range_data, grid)
}
\arguments{
\item{range_data}{Data frame of species ranges}

\item{grid}{Grid data frame for spatial matching}
}
\value{
A list of matched ranges
}
\description{
Prepare range data to match the grid
}
